+++ packages/extension/src/companion/CompanionDiscussion.tsx
@@ -4,7 +4,7 @@ import { PostBootData } from '@dailydotdev/shared/src/lib/boot';
 import { NewComment } from '@dailydotdev/shared/src/components/post/NewComment';
 import { PostComments } from '@dailydotdev/shared/src/components/post/PostComments';
 import AuthContext from '@dailydotdev/shared/src/contexts/AuthContext';
-import NewCommentModal from '@dailydotdev/shared/src/components/modals/NewCommentModal';
+import NewCommentModal from '@dailydotdev/shared/src/components/modals/comment/NewCommentModal';
 import ShareModal from '@dailydotdev/shared/src/components/modals/ShareModal';
 import { Origin } from '@dailydotdev/shared/src/lib/analytics';
 import { useShareComment } from '@dailydotdev/shared/src/hooks/useShareComment';
